refactor

+++ 1Jirafapi/src/main/java/scanpackage/models/StatusModel.java
@@ -3,15 +3,15 @@ package scanpackage.models;
 import java.util.Objects;
 
 public class StatusModel {
-    private long id;
+    private int id;
 
     private String name;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -49,7 +49,7 @@ public class StatusModel {
 
     }
 
-    public StatusModel(long id, String name) {
+    public StatusModel(int id, String name) {
         this.id = id;
         this.name = name;
     }
